Add system & primary clipboards as special registers
These special registers join and copy the values to the clipboards
with '*' corresponding to the system clipboard and '+' to the primary
as they are in Vim. This also uses the trick from PR6889 to save the
values in the register and re-use them without joining into one
value when pasting a value which was yanked and not changed.

These registers are not implemented in Kakoune but Kakoune also does
not have built-in clipboard integration.

Co-authored-by: CcydtN <51289140+CcydtN@users.noreply.github.com>
Co-authored-by: Pascal Kuthe <pascal.kuthe@semimod.de>